+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.02.01',
+  '2023.02.07',
   // Fallback runtime versions.
+  '2023.02.01',
   '2023.01.25',
-  '2023.01.24',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -54,10 +54,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.19.0';
-final String toolStableDartSdkVersion = '2.19.1';
-final String toolStableFlutterSdkVersion = '3.7.0';
-final String toolPreviewDartSdkVersion = '3.0.0-179.0.dev';
-final String toolPreviewFlutterSdkVersion = '3.7.0';
+final String toolStableDartSdkVersion = '2.19.2';
+final String toolStableFlutterSdkVersion = '3.7.1';
+final String toolPreviewDartSdkVersion = '3.0.0-187.0.dev';
+final String toolPreviewFlutterSdkVersion = '3.7.1';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);